requirements.txt
langchain==0.1.1
openai==1.8.0
pydantic==2.5.3
python-dotenv==1.0.0
streamlit==1.26.0
tenacity==8.2.2
